refactor file descriptors
