+++ prowler/lib/outputs/outputs.py
@@ -1,31 +1,24 @@
 import json
 import sys
-from csv import DictWriter
-from io import TextIOWrapper
-from typing import Any
 
 from colorama import Fore, Style
 
 from prowler.config.config import (
     csv_file_suffix,
-    html_file_suffix,
     json_asff_file_suffix,
     json_file_suffix,
     orange_color,
-    timestamp,
 )
 from prowler.lib.logger import logger
-from prowler.lib.outputs.csv import generate_csv_fields
-from prowler.lib.outputs.html import add_html_header, fill_html
+from prowler.lib.outputs.compliance import fill_compliance
+from prowler.lib.outputs.file_descriptors import fill_file_descriptors
+from prowler.lib.outputs.html import fill_html
 from prowler.lib.outputs.json import fill_json_asff
 from prowler.lib.outputs.models import (
-    Aws_Check_Output_CSV,
-    Azure_Check_Output_CSV,
-    Check_Output_CSV_CIS,
-    Check_Output_CSV_ENS_RD2022,
     Check_Output_JSON_ASFF,
+    generate_provider_output_csv,
+    generate_provider_output_json,
 )
-from prowler.lib.utils.utils import file_exists, open_file
 from prowler.providers.aws.lib.allowlist.allowlist import is_allowlisted
 from prowler.providers.aws.lib.audit_info.models import AWS_Audit_Info
 from prowler.providers.aws.lib.security_hub.security_hub import send_to_security_hub
@@ -90,137 +83,17 @@ def report(check_findings, output_options, audit_info):
                 if file_descriptors:
                     # AWS specific outputs
                     if finding.check_metadata.Provider == "aws":
-                        if "ens_rd2022_aws" in output_options.output_modes:
-                            # We have to retrieve all the check's compliance requirements
-                            check_compliance = output_options.bulk_checks_metadata[
-                                finding.check_metadata.CheckID
-                            ].Compliance
-                            for compliance in check_compliance:
-                                if (
-                                    compliance.Framework == "ENS"
-                                    and compliance.Version == "RD2022"
-                                ):
-                                    for requirement in compliance.Requirements:
-                                        requirement_description = (
-                                            requirement.Description
-                                        )
-                                        requirement_id = requirement.Id
-                                        for attribute in requirement.Attributes:
-                                            compliance_row = Check_Output_CSV_ENS_RD2022(
-                                                Provider=finding.check_metadata.Provider,
-                                                AccountId=audit_info.audited_account,
-                                                Region=finding.region,
-                                                AssessmentDate=timestamp.isoformat(),
-                                                Requirements_Id=requirement_id,
-                                                Requirements_Description=requirement_description,
-                                                Requirements_Attributes_IdGrupoControl=attribute.get(
-                                                    "IdGrupoControl"
-                                                ),
-                                                Requirements_Attributes_Marco=attribute.get(
-                                                    "Marco"
-                                                ),
-                                                Requirements_Attributes_Categoria=attribute.get(
-                                                    "Categoria"
-                                                ),
-                                                Requirements_Attributes_DescripcionControl=attribute.get(
-                                                    "DescripcionControl"
-                                                ),
-                                                Requirements_Attributes_Nivel=attribute.get(
-                                                    "Nivel"
-                                                ),
-                                                Requirements_Attributes_Tipo=attribute.get(
-                                                    "Tipo"
-                                                ),
-                                                Requirements_Attributes_Dimensiones=",".join(
-                                                    attribute.get("Dimensiones")
-                                                ),
-                                                Status=finding.status,
-                                                StatusExtended=finding.status_extended,
-                                                ResourceId=finding.resource_id,
-                                                CheckId=finding.check_metadata.CheckID,
-                                            )
-
-                                    csv_header = generate_csv_fields(
-                                        Check_Output_CSV_ENS_RD2022
-                                    )
-                                    csv_writer = DictWriter(
-                                        file_descriptors["ens_rd2022_aws"],
-                                        fieldnames=csv_header,
-                                        delimiter=";",
-                                    )
-                                    csv_writer.writerow(compliance_row.__dict__)
-                        elif "cis" in str(output_options.output_modes):
-                            # We have to retrieve all the check's compliance requirements
-                            check_compliance = output_options.bulk_checks_metadata[
-                                finding.check_metadata.CheckID
-                            ].Compliance
-                            for compliance in check_compliance:
-                                if compliance.Framework == "CIS-AWS":
-                                    for requirement in compliance.Requirements:
-                                        requirement_description = (
-                                            requirement.Description
-                                        )
-                                        requirement_id = requirement.Id
-                                        for attribute in requirement.Attributes:
-                                            compliance_row = Check_Output_CSV_CIS(
-                                                Provider=finding.check_metadata.Provider,
-                                                AccountId=audit_info.audited_account,
-                                                Region=finding.region,
-                                                AssessmentDate=timestamp.isoformat(),
-                                                Requirements_Id=requirement_id,
-                                                Requirements_Description=requirement_description,
-                                                Requirements_Attributes_Section=attribute.get(
-                                                    "Section"
-                                                ),
-                                                Requirements_Attributes_Profile=attribute.get(
-                                                    "Profile"
-                                                ),
-                                                Requirements_Attributes_AssessmentStatus=attribute.get(
-                                                    "AssessmentStatus"
-                                                ),
-                                                Requirements_Attributes_Description=attribute.get(
-                                                    "Description"
-                                                ),
-                                                Requirements_Attributes_RationaleStatement=attribute.get(
-                                                    "RationaleStatement"
-                                                ),
-                                                Requirements_Attributes_ImpactStatement=attribute.get(
-                                                    "ImpactStatement"
-                                                ),
-                                                Requirements_Attributes_RemediationProcedure=attribute.get(
-                                                    "RemediationProcedure"
-                                                ),
-                                                Requirements_Attributes_AuditProcedure=attribute.get(
-                                                    "AuditProcedure"
-                                                ),
-                                                Requirements_Attributes_AdditionalInformation=attribute.get(
-                                                    "AdditionalInformation"
-                                                ),
-                                                Requirements_Attributes_References=attribute.get(
-                                                    "References"
-                                                ),
-                                                Status=finding.status,
-                                                StatusExtended=finding.status_extended,
-                                                ResourceId=finding.resource_id,
-                                                CheckId=finding.check_metadata.CheckID,
-                                            )
-
-                                    csv_header = generate_csv_fields(
-                                        Check_Output_CSV_CIS
-                                    )
-                                    csv_writer = DictWriter(
-                                        file_descriptors[
-                                            output_options.output_modes[-1]
-                                        ],
-                                        fieldnames=csv_header,
-                                        delimiter=";",
-                                    )
-                                    csv_writer.writerow(compliance_row.__dict__)
+                        if (
+                            "ens_rd2022_aws" in output_options.output_modes
+                            or "cis" in str(output_options.output_modes)
+                        ):
+                            fill_compliance(
+                                output_options, finding, audit_info, file_descriptors
+                            )
 
                         if "html" in file_descriptors:
-                            fill_html(file_descriptors["html"], finding)
-
-                        file_descriptors["html"].write("")
+                            fill_html(file_descriptors["html"], audit_info, finding)
+                            file_descriptors["html"].write("")
 
                         if "json-asff" in file_descriptors:
                             finding_output = Check_Output_JSON_ASFF()
@@ -280,124 +153,6 @@ def report(check_findings, output_options, audit_info):
         logger.error(
             f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
         )
-
-
-def initialize_file_descriptor(
-    filename: str,
-    output_mode: str,
-    audit_info: AWS_Audit_Info,
-    format: Any = None,
-) -> TextIOWrapper:
-    """Open/Create the output file. If needed include headers or the required format"""
-    try:
-        if file_exists(filename):
-            file_descriptor = open_file(
-                filename,
-                "a",
-            )
-        else:
-            file_descriptor = open_file(
-                filename,
-                "a",
-            )
-
-            if output_mode in ("csv", "ens_rd2022_aws", "cis_1.5_aws", "cis_1.4_aws"):
-                # Format is the class model of the CSV format to print the headers
-                csv_header = [x.upper() for x in generate_csv_fields(format)]
-                csv_writer = DictWriter(
-                    file_descriptor, fieldnames=csv_header, delimiter=";"
-                )
-                csv_writer.writeheader()
-
-            if output_mode in ("json", "json-asff"):
-                file_descriptor.write("[")
-            if "html" in output_mode:
-                add_html_header(file_descriptor, audit_info)
-    except Exception as error:
-        logger.error(
-            f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
-        )
-
-    return file_descriptor
-
-
-def fill_file_descriptors(output_modes, output_directory, output_filename, audit_info):
-    try:
-        file_descriptors = {}
-        if output_modes:
-            for output_mode in output_modes:
-                if output_mode == "csv":
-                    filename = f"{output_directory}/{output_filename}{csv_file_suffix}"
-                    if isinstance(audit_info, AWS_Audit_Info):
-                        file_descriptor = initialize_file_descriptor(
-                            filename,
-                            output_mode,
-                            audit_info,
-                            Aws_Check_Output_CSV,
-                        )
-                    if isinstance(audit_info, Azure_Audit_Info):
-                        file_descriptor = initialize_file_descriptor(
-                            filename,
-                            output_mode,
-                            audit_info,
-                            Azure_Check_Output_CSV,
-                        )
-                    file_descriptors.update({output_mode: file_descriptor})
-
-                if output_mode == "json":
-                    filename = f"{output_directory}/{output_filename}{json_file_suffix}"
-                    file_descriptor = initialize_file_descriptor(
-                        filename, output_mode, audit_info
-                    )
-                    file_descriptors.update({output_mode: file_descriptor})
-
-                if isinstance(audit_info, AWS_Audit_Info):
-
-                    if output_mode == "json-asff":
-                        filename = f"{output_directory}/{output_filename}{json_asff_file_suffix}"
-                        file_descriptor = initialize_file_descriptor(
-                            filename, output_mode, audit_info
-                        )
-                        file_descriptors.update({output_mode: file_descriptor})
-
-                    if output_mode == "html":
-                        filename = (
-                            f"{output_directory}/{output_filename}{html_file_suffix}"
-                        )
-                        file_descriptor = initialize_file_descriptor(
-                            filename, output_mode, audit_info
-                        )
-                        file_descriptors.update({output_mode: file_descriptor})
-
-                    if output_mode == "ens_rd2022_aws":
-                        filename = f"{output_directory}/{output_filename}_ens_rd2022_aws{csv_file_suffix}"
-                        file_descriptor = initialize_file_descriptor(
-                            filename,
-                            output_mode,
-                            audit_info,
-                            Check_Output_CSV_ENS_RD2022,
-                        )
-                        file_descriptors.update({output_mode: file_descriptor})
-
-                    if output_mode == "cis_1.5_aws":
-                        filename = f"{output_directory}/{output_filename}_cis_1.5_aws{csv_file_suffix}"
-                        file_descriptor = initialize_file_descriptor(
-                            filename, output_mode, audit_info, Check_Output_CSV_CIS
-                        )
-                        file_descriptors.update({output_mode: file_descriptor})
-
-                    if output_mode == "cis_1.4_aws":
-                        filename = f"{output_directory}/{output_filename}_cis_1.4_aws{csv_file_suffix}"
-                        file_descriptor = initialize_file_descriptor(
-                            filename, output_mode, audit_info, Check_Output_CSV_CIS
-                        )
-                        file_descriptors.update({output_mode: file_descriptor})
-    except Exception as error:
-        logger.error(
-            f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
-        )
-
-    return file_descriptors
 
 
 def set_report_color(status: str) -> str:
